bk: 472ba5afzVb3Kzb1huLxuFXH5zwC2g

+++ generic/tclExecute.c
@@ -2413,7 +2413,7 @@ TclExecuteByteCode(
      *	   Start of INST_LOAD instructions.
      *
      * WARNING: more 'goto' here than your doctor recommended! The different
-     * instructions set the value of some variables and then jump to somme
+     * instructions set the value of some variables and then jump to some
      * common execution code.
      */
     {
@@ -5493,7 +5493,8 @@ TclExecuteByteCode(
 		if (l1 >= 3
 		    && (unsigned long) l1 < (sizeof(Exp32Index)
 			     / sizeof(unsigned short)) - 1) {
-		    unsigned short base = Exp32Index[l1-3] + l2 - 9;
+		    unsigned short base = Exp32Index[l1-3]
+			                + (unsigned short) l2 - 9;
 		    if (base < Exp32Index[l1-2]) {
 			/*
 			 * 32-bit number raised to intermediate power,
@@ -5513,7 +5514,8 @@ TclExecuteByteCode(
 		if (-l1 >= 3
 		    && (unsigned long)(-l1) < (sizeof(Exp32Index)
 			     / sizeof(unsigned short)) - 1) {
-		    unsigned short base = Exp32Index[-l1-3] + l2 - 9;
+		    unsigned short base
+			= Exp32Index[-l1-3] + (unsigned short) l2 - 9;
 		    if (base < Exp32Index[-l1-2]) {
 			long lResult = (oddExponent) ? 
 			    -Exp32Value[base] : Exp32Value[base];
@@ -5637,7 +5639,8 @@ TclExecuteByteCode(
 	    if (w1 >= 3
 		&& (Tcl_WideUInt) w1 < (sizeof(Exp64Index)
 			 / sizeof(unsigned short)) - 1) {
-		unsigned short base = Exp64Index[w1-3] + l2 - 17;
+		unsigned short base
+		    = Exp64Index[w1-3] + (unsigned short) l2 - 17;
 		if (base < Exp64Index[w1-2]) {
 		    /*
 		     * 64-bit number raised to intermediate power,
@@ -5657,7 +5660,8 @@ TclExecuteByteCode(
 	    if (-w1 >= 3
 		&& (Tcl_WideUInt) (-w1) < (sizeof(Exp64Index)
 			  / sizeof(unsigned short)) - 1) {
-		unsigned short base = Exp64Index[-w1-3] + l2 - 17;
+		unsigned short base
+		    = Exp64Index[-w1-3] + (unsigned short) l2 - 17;
 		if (base < Exp64Index[-w1-2]) {
 		    Tcl_WideInt wResult = (oddExponent) ? 
 			-Exp64Value[base] : Exp64Value[base];
